KAS-1218: indicatie BIS document bij workflow kb
