+++ app/components/agenda/agendaitem/agendaitem-overview/component.js
@@ -4,13 +4,16 @@ import {computed, observer} from '@ember/object';
 import {alias} from '@ember/object/computed';
 import isAuthenticatedMixin from 'fe-redpencil/mixins/is-authenticated-mixin';
 import DS from 'ember-data';
+import { tracked } from '@glimmer/tracking';
 
 export default Component.extend(isAuthenticatedMixin, {
   classNames: ['vlc-panel-layout__main-content'],
   currentAgenda: alias('sessionService.currentAgenda'),
   sessionService: inject(),
   store: inject(),
+  agendaService: inject(),
   activeAgendaItemSection: 'details',
+  @tracked showNewsItemIsEditedWarning:false,
 
   checkAgendaItemSubcase: observer('subcase', function () {
     this.get('subcase').then((subcase) => {
@@ -49,7 +52,11 @@ export default Component.extend(isAuthenticatedMixin, {
       }
     },
 
-    setAgendaItemSection(section) {
+    async setAgendaItemSection(section) {
+      if(section==='news-item') {
+        const subcase = await this.get('subcase');
+        this.showNewsItemIsEditedWarning = await this.agendaService.shouldShowEditedWarning(subcase);
+      }
       this.set('activeAgendaItemSection', section);
     },
