+++ app/components/agenda/agendaitem/agendaitem-overview/template.hbs
@@ -68,6 +68,7 @@
             {{agenda/agendaitem/agendaitem-news-item
               agendaitem=agendaitem
               subcase=subcase
+              showNewsItemIsEditedWarning=showNewsItemIsEditedWarning
             }}
           {{else if (eq activeAgendaItemSection "minutes")}}
             {{agenda/agendaitem/agendaitem-minutes agendaitem=agendaitem}}
