+++ app/services/session-service.js
@@ -8,7 +8,7 @@ export default Service.extend({
   router: inject(),
   currentSession: null,
   selectedAgendaItem: null,
-  showNewsItemIsEditedWarning: null,
+  showNewsItemIsEditedWarning: false,
 
   agendas: computed('currentSession.agendas.@each', function () {
     if (!this.get('currentSession')) {
